changed map to be a button to choose between dark mode and light mode
