+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "worther",
-  "version": "1.0.4",
+  "version": "1.0.5",
   "private": true,
   "dependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
