+++ src/pages/app.jsx
@@ -2,7 +2,6 @@ import '../styles/main.scss';
 import Home from './home.jsx';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import About from './about.jsx';
-import Map from './map.jsx';
 import Weather from './weather.jsx';
 import AdvancedWeather from './advancedWeather.jsx';
 import { GetOpenWeatherData } from '../components/openWeatherData.jsx';
@@ -21,7 +20,6 @@ export default function App() {
         <Routes>
           <Route exact path="/" element={<Home />} />
           <Route exact path="/about" element={<About />} />
-          <Route exact path="/map" element={<Map />} />
           <Route exact path="/map/:mapType" element={<ShowMap />} />
           <Route exact path="/weather" element={<Weather />} />
           <Route exact path="/advancedWeather" element={<AdvancedWeather />} />
